chore(controllers) remove extra business logics from the controllers [FINISHES #167216160]
