+++ src/controllers/signin.js
@@ -1,6 +1,6 @@
 import bcrypt from 'bcrypt';
 import { results, SUCCESS, ERROR } from '../helper/result';
-import { generateAuthToken, validateLogin, validateEmail } from '../helper/user';
+import { generateAuthToken, validateLogin, validateEmail, validatePassword } from '../helper/user';
 
 const signin = async (req, res) => {
   const { error } = validateLogin(req);
@@ -9,7 +9,7 @@ const signin = async (req, res) => {
   const user = validateEmail(req);
   if (!user) return res.status(400).send(results(400, ERROR, 'Invalid email'));
 
-  const isValid = await bcrypt.compare(req.body.password, user.password);
+  const isValid = await validatePassword(req, user);
   if (!isValid) return res.status(400).send(results(400, ERROR, 'Invalid password'));
 
   const token = generateAuthToken(user);
