+++ src/controllers/signup.js
@@ -1,27 +1,18 @@
 import bcrypt from 'bcrypt';
 import _ from 'lodash';
-import { users, validateSignup, generateAuthToken } from '../helper/user';
 import { results, SUCCESS, ERROR } from '../helper/result';
+import { users, validateSignup, generateAuthToken, validateEmail, save, validatePassword, hashPassword } from '../helper/user';
 
 const signup = async (req, res) => {
   const { error } = validateSignup(req);
   if (error) return res.status(400).send(results(400, ERROR, error.details[0].message));
 
-  let user = users.find(user => user.email === req.body.email);
+  let user = validateEmail(req);
   if (user) return res.status(400).send(results(400, ERROR, 'user already registered.'));
 
-  user = _.pick(req.body, ['first_name', 'last_name', 'email', 'password', 'phoneNumber', 'address', 'isAdmin']);
-  user.id = users.length + 1;
+  user = await save(req);
 
-  const token = generateAuthToken(user);
-  user.token = token;
-
-  const salt = await bcrypt.genSalt(10);
-  user.password = await bcrypt.hash(user.password, salt);
-
-  users.push(user);
-
-  res.header('x-auth-token', token).send(results(200, SUCCESS, user));
+  res.header('x-auth-token', user.token).send(results(200, SUCCESS, user));
 };
 
 export default signup;
